Add BERT FT Eval script
