--- usc26/FTBERT_HPO_Results/results_lr2e-5_bs16_ep4.json
@@ -1,17 +1,0 @@
-{
-  "job_id": "lr2e-5_bs16_ep4",
-  "results": [
-    {
-      "lr": 2e-05,
-      "batch_size": 16,
-      "epochs": 4,
-      "macro_f1": 0.6831394077790923
-    }
-  ],
-  "best": {
-    "lr": 2e-05,
-    "batch_size": 16,
-    "epochs": 4,
-    "macro_f1": 0.6831394077790923
-  }
-}
