--- usc26/FTBERT_HPO_Results/results_lr3e-5_bs16_ep3.json
@@ -1,17 +1,0 @@
-{
-  "job_id": "lr3e-5_bs16_ep3",
-  "results": [
-    {
-      "lr": 3e-05,
-      "batch_size": 16,
-      "epochs": 3,
-      "macro_f1": 0.6759375537520531
-    }
-  ],
-  "best": {
-    "lr": 3e-05,
-    "batch_size": 16,
-    "epochs": 3,
-    "macro_f1": 0.6759375537520531
-  }
-}
