--- usc26/FTBERT_HPO_Results/results_lr3e-5_bs8_ep4.json
@@ -1,17 +1,0 @@
-{
-  "job_id": "lr3e-5_bs8_ep4",
-  "results": [
-    {
-      "lr": 3e-05,
-      "batch_size": 8,
-      "epochs": 4,
-      "macro_f1": 0.8249784800452273
-    }
-  ],
-  "best": {
-    "lr": 3e-05,
-    "batch_size": 8,
-    "epochs": 4,
-    "macro_f1": 0.8249784800452273
-  }
-}
