+++ usc26/FineTuneBERT_USC26/FTBERT_HPO_Results/results_lr3e-5_bs8_ep3.json
@@ -1,0 +1,17 @@
+{
+  "job_id": "lr3e-5_bs8_ep3",
+  "results": [
+    {
+      "lr": 3e-05,
+      "batch_size": 8,
+      "epochs": 3,
+      "macro_f1": 0.6792165363902842
+    }
+  ],
+  "best": {
+    "lr": 3e-05,
+    "batch_size": 8,
+    "epochs": 3,
+    "macro_f1": 0.6792165363902842
+  }
+}
