+++ usc26/FineTuneBERT_USC26/FTBERT_HPO_Results/results_test.json
@@ -1,0 +1,17 @@
+{
+  "job_id": "test",
+  "results": [
+    {
+      "lr": 2e-05,
+      "batch_size": 8,
+      "epochs": 5,
+      "macro_f1": 0.8152005166214167
+    }
+  ],
+  "best": {
+    "lr": 2e-05,
+    "batch_size": 8,
+    "epochs": 5,
+    "macro_f1": 0.8152005166214167
+  }
+}
